requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
click-completion==0.5.2
click-default-group==1.2.2
colorama==0.4.4
commonmark==0.9.1
cucumber-tag-expressions==2.0.4
flake8==3.9.2
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.1.0
pluggy==0.13.1
pprintpp==0.4.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
python-dotenv==0.17.1
rich==10.2.2
shellingham==1.4.0
six==1.16.0
toml==0.10.2
twitchio==1.2.3
typing-extensions==3.10.0.0
ward==0.59.0b0
websockets==9.1
yarl==1.6.3
